+++ .gitignore
@@ -1,3 +1,4 @@
+.base/*
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
