+++ mmrazor/models/architectures/dynamic_ops/mixins/__init__.py
@@ -1,16 +1,14 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .dynamic_conv_mixins import DynamicConvMixin
 from .dynamic_layernorm_mixins import DynamicLayerNormMixin
-from .dynamic_mha_mixins import DynamicMHAMixin
 from .dynamic_mixins import (DynamicBatchNormMixin, DynamicChannelMixin,
                              DynamicLinearMixin, DynamicMixin)
-from .dynamic_patchembed_mixins import DynamicPatchEmbedMixin
-from .dynamic_rp_mixins import DynamicRelativePosition2DMixin
-from .dynamic_squential_mixins import DynamicSequentialMixin
 
 __all__ = [
-    'DynamicChannelMixin', 'DynamicBatchNormMixin', 'DynamicLinearMixin',
-    'DynamicMixin', 'DynamicConvMixin', 'DynamicSequentialMixin',
-    'DynamicPatchEmbedMixin', 'DynamicLayerNormMixin',
-    'DynamicRelativePosition2DMixin', 'DynamicMHAMixin'
+    'DynamicChannelMixin',
+    'DynamicBatchNormMixin',
+    'DynamicLinearMixin',
+    'DynamicMixin',
+    'DynamicConvMixin',
+    'DynamicLayerNormMixin',
 ]
